feat: a lot of new things

+++ product/productServices.js
@@ -1,4 +1,6 @@
 const Product = require("./productModel");
+const jwt = require("jsonwebtoken");
+const adminRoleModel = require("../admin/adminRoleModel");
 
 const createProduct = async (productData) => {
   try {
@@ -101,10 +103,35 @@ const updateProduct = async (productData, productId) => {
   }
 };
 
-const deleteProduct = async (productId) => {
+const deleteProduct = async (req) => {
   try {
-    await Product.findByIdAndDelete(productId);
-    return "Producto eliminado exitosamente";
+    const adminToken = req.body.adminToken;
+    const productId = req.params.id;
+    let check;
+    jwt.verify(adminToken, process.env.JWT_SECRET, async (err, decoded) => {
+      let result = await adminRoleModel.findOne({
+        area: decoded.area,
+      });
+      check = result;
+      if (err) {
+        return res.status(500).send({
+          auth: false,
+          message: "Falló autenticación de token.",
+        });
+      } else if (decoded) {
+        check = result;
+        if (check && check.deleteProduct) {
+          await Product.findByIdAndDelete(productId);
+          return "Producto eliminado exitosamente";
+        } else {
+          const warning = {
+            auth: false,
+            message: "No tienes autorización para realizar esta acción.",
+          };
+          return warning;
+        }
+      }
+    });
   } catch (error) {
     console.log(error);
     return error;
